Move mock files to subdirs mirroring the contract layout
