+++ contracts/mocks/token/ERC7984/ERC7984Mock.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.27;
 
 import {ZamaEthereumConfig} from "@fhevm/solidity/config/ZamaConfig.sol";
 import {FHE, eaddress, euint64, externalEuint64} from "@fhevm/solidity/lib/FHE.sol";
-import {ERC7984} from "../../token/ERC7984/ERC7984.sol";
+import {ERC7984} from "../../../token/ERC7984/ERC7984.sol";
 
 // solhint-disable func-name-mixedcase
 contract ERC7984Mock is ERC7984, ZamaEthereumConfig {
