+++ contracts/mocks/token/ERC7984/extensions/ERC7984ObserverAccessMock.sol
@@ -3,8 +3,8 @@
 pragma solidity ^0.8.27;
 
 import {euint64} from "@fhevm/solidity/lib/FHE.sol";
-import {ERC7984ObserverAccess} from "../../token/ERC7984/extensions/ERC7984ObserverAccess.sol";
-import {ERC7984Mock} from "./ERC7984Mock.sol";
+import {ERC7984ObserverAccess} from "../../../../token/ERC7984/extensions/ERC7984ObserverAccess.sol";
+import {ERC7984Mock} from "../ERC7984Mock.sol";
 
 contract ERC7984ObserverAccessMock is ERC7984ObserverAccess, ERC7984Mock {
     constructor(
