+++ contracts/mocks/token/ERC7984/extensions/ERC7984RestrictedMock.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.27;
 
 import {ZamaEthereumConfig} from "@fhevm/solidity/config/ZamaConfig.sol";
 import {FHE} from "@fhevm/solidity/lib/FHE.sol";
-import {ERC7984Restricted} from "../../token/ERC7984/extensions/ERC7984Restricted.sol";
+import {ERC7984Restricted} from "../../../../token/ERC7984/extensions/ERC7984Restricted.sol";
 
 abstract contract ERC7984RestrictedMock is ERC7984Restricted, ZamaEthereumConfig {
     function _mint(address to, uint64 amount) internal {
